requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
redis==5.2.1
requests==2.32.3
python-dotenv==1.0.1
